requirements.txt
# CICosts API Dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
mangum==0.17.0  # Lambda adapter

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
httpx==0.26.0

# AWS
boto3==1.34.25

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-multipart==0.0.6

# Monitoring (optional)
sentry-sdk[fastapi]==1.39.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
